footer수정

+++ .eslintrc.js
@@ -13,7 +13,7 @@ module.exports = {
   rules: {
     'no-console': process.env.NODE_ENV === 'production' ? 'warn' : 'off',
     'no-debugger': process.env.NODE_ENV === 'production' ? 'warn' : 'off',
-    "quotes": [2, "double", { "avoidEscape": false }],
+    "quotes": [2, "single", { "avoidEscape": false }],
   }
 
 }
